Move types to it's own file to allow for better import into GoogleAppsScripts
