+++ src/app/cap-table/state/ConversionState.ts
@@ -4,7 +4,7 @@ import { CurrencyInputOnChangeValues } from "react-currency-input-field";
 import { SeriesProps } from "@/components/safe-conversion/Conversion/SeriesInvestorList";
 import { SAFEProps } from "@/components/safe-conversion/Conversion/SafeNoteList";
 import { ExistingShareholderProps } from "@/components/safe-conversion/Conversion/ExistingShareholders";
-import { CapTableRowType, CommonRowType } from "@library/cap-table";
+import { CapTableRowType, CommonRowType } from "@library/cap-table/types";
 
 // Only the state that we need to serialize
 export type ExistingShareholderState = Pick<
