+++ src/app/cap-table/state/selectors/PricedRoundSelector.ts
@@ -1,6 +1,6 @@
 import { createSelector } from "reselect";
 import { ExistingShareholderState, IConversionStateData, IRowState, SAFEState } from "../ConversionState";
-import { buildPricedRoundCapTable, CapTableRowType, CommonRowType, CommonStockholder, SAFENote, SeriesInvestor } from "@library/cap-table";
+import { buildPricedRoundCapTable } from "@library/cap-table";
 import { CapTableProps } from "@/components/safe-conversion/Conversion/CapTableResults";
 import { BestFit, fitConversion } from "@library/conversion-solver";
 import { stringToNumber } from "@library/utils/numberFormatting";
@@ -8,6 +8,7 @@ import { SAFEProps } from "@/components/safe-conversion/Conversion/SafeNoteList"
 import { getCapForSafe } from "@library/safe-calcs";
 import { SeriesProps } from "@/components/safe-conversion/Conversion/SeriesInvestorList";
 import { PricedRoundPropsData } from "@/components/safe-conversion/Conversion/PricedRound";
+import { CommonStockholder, SAFENote, SeriesInvestor, CapTableRowType, CommonRowType } from "@library/cap-table/types";
 
 export type ResultSelectorState = IConversionStateData & {
   preMoneyChange?: number;
